requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
deepface>=0.0.93
tensorflow>=2.16.0
opencv-python-headless>=4.9.0.80
numpy>=1.26.0
Pillow>=10.3.0
aiofiles>=23.2.1
